+++ changelog.md
@@ -36,6 +36,7 @@ Removed effects:
 - Hunter's Arcane Shot no longer glows during Lock and Load (Cataclysm)
 - Hunter's Improved Steady Shot has been removed (Cataclysm)
 - Rogue's Riposte has been removed (Cataclysm)
+- Warlock's Soul Fire no longer glows during Molten Core (Cataclysm)
 
 Sounds are no longer played for minor alerts, including:
 - Mage's Clearcasting